初步完成方块属性加载（属性文件需要更新，大概得研究下怎么从bds拿:(）

+++ Allay-API/src/main/java/cn/allay/block/component/BlockDataComponent.java
@@ -30,10 +30,10 @@ public interface BlockDataComponent {
     boolean fallable();
     float friction();
     float hardness();
-    int resistance();
+    float resistance();
     boolean solid();
     boolean pushesOutItems();
-    int translucency();
+    float translucency();
     int burnChance();
     int burnAbility();
     int lightFilterFactor();
